moved GrandIdEnvironmentConfiguration to constructor and changed comments
