+++ src/ActiveLogin.Authentication.GrandId.Api/GrandIdHttpClientExtensions.cs
@@ -1,5 +1,4 @@
 ﻿using System.Net.Http;
-using System.Text;
 using System.Threading.Tasks;
 using ActiveLogin.Authentication.Common.Serialization;
