+++ src/ActiveLogin.Authentication.GrandId.Api/GrandIdUrls.cs
@@ -3,17 +3,17 @@
 namespace ActiveLogin.Authentication.GrandId.Api
 {
     /// <summary>
-    /// URLs for BankID REST API.
+    /// URLs for GrandID REST API.
     /// </summary>
     public static class GrandIdUrls
     {
         /// <summary>
-        /// Base url for production API. Needs to be used in conjunction with a production certificate.
+        /// Base url for production API.
         /// </summary>
         public static readonly Uri ProdApiBaseUrl = new Uri("https://client.grandid.com/json1.1/");
 
         /// <summary>
-        /// Base url for test API. Needs to be used in conjunction with the test certificate.
+        /// Base url for test API.
         /// </summary>
         public static readonly Uri TestApiBaseUrl = new Uri("https://client-test.grandid.com/json1.1/");
     }
