+++ src/ActiveLogin.Authentication.GrandId.Api/IGrandIdEnviromentConfiguration.cs
@@ -1,7 +1,7 @@
-﻿using ActiveLogin.Authentication.GrandId.Api.Models;
-using System;
+﻿using System;
+using ActiveLogin.Authentication.GrandId.Api.Models;
 
-namespace ActiveLogin.Authentication.GrandId
+namespace ActiveLogin.Authentication.GrandId.Api
 {
     public interface IGrandIdEnviromentConfiguration
     {
