+++ src/ActiveLogin.Authentication.GrandId.Api/Models/SessionStateResponse.cs
@@ -1,6 +1,6 @@
 ﻿using System.Runtime.Serialization;
 
-namespace ActiveLogin.Authentication.GrandId.Api
+namespace ActiveLogin.Authentication.GrandId.Api.Models
 {
     [DataContract]
     public class SessionStateResponse
